UFO Crash Scene - Final Commit
This is the Final commit of the UFO crash Scene. I made this scene in unity in the standard 3D render pipeline. The UFO model at the centre of the scene is custom made in blender. All the other models were downloaded either from the AssetStore or from various other open source sites. I do not take credit for those models. However the scene as a whole has been made by me.